+++ include/inc/hooks.h
@@ -26,20 +26,20 @@
 # define ZOOM_OUT		4
 # define ZOOM_IN		5
 
-/*
 # define ROTATE_LEFT 65361
 # define ROTATE_RIGHT 65363
 # define LEFT	97
 # define UP		119
 # define RIGHT	100
 # define DOWN	115
-*/
 
+/*
 # define ROTATE_LEFT 123
 # define ROTATE_RIGHT 124
 # define LEFT	0
 # define UP		13
 # define RIGHT	2
 # define DOWN	1
+*/
 
 #endif
