+++ src/parser/player.c
@@ -17,13 +17,17 @@ int process_player(int pos_found, int line, int colum, char **raw)
     int i;
     int x;
 
-    if (pos_found => 1)
+    (void)x;
+    (void)line;
+    (void)colum;
+    (void)raw;
+    if (pos_found >= 1)
         return (-1);
     i = 0;
     while (raw[i])
     {
         x = 0;
-        while (raw)
+        //while (raw)
         i++;
     }
