+++ src/render/ray.c
@@ -1,0 +1,55 @@
+#include <math.h>
+
+#include "cub3d.h"
+
+static void	ray_coord(t_ray *ray, t_player *pl, t_camera *cam)
+{
+	ray->dir_x = pl->dir_x + cam->plane_x * cam->coord_x;
+	ray->dir_y = pl->dir_y + cam->plane_y * cam->coord_x;
+	ray->map_x = (int)ray->pos_x;
+	ray->map_y = (int)ray->pos_y;
+}
+
+static void	ray_next_square_dist(t_ray *ray)
+{
+	if (ray->dir_x < 0)
+	{
+		ray->step_x = -1;
+		ray->side_dist_x = (ray->pos_x - ray->map_x) * ray->delta_dist_x;
+	}
+	else
+	{
+		ray->step_x = 1;
+		ray->side_dist_x = (ray->map_x + 1.0 - ray->pos_x) * ray->delta_dist_x;
+	}
+	if (ray->dir_y < 0)
+	{
+		ray->step_y = -1;
+		ray->side_dist_y = (ray->pos_y - ray->map_y) * ray->delta_dist_y;
+	}
+	else
+	{
+		ray->step_y = 1;
+		ray->side_dist_y = (ray->map_y + 1.0 - ray->pos_y) * ray->delta_dist_y;
+	}
+}
+
+static void	ray_delta(t_ray *ray)
+{
+	if (!ray->dir_x)
+		ray->delta_dist_x = 1e30;
+	else
+		ray->delta_dist_x = fabs(1 / ray->dir_x);
+	if (!ray->dir_y)
+		ray->delta_dist_y = 1e30;
+	else
+		ray->delta_dist_y = fabs(1 / ray->dir_y);
+}
+
+void	set_ray(t_ray *ray, t_player *pl, t_camera *cam)
+{
+	ray_coord(ray, pl, cam);
+	ray_delta(ray);
+	ray_next_square_dist(ray);
+}
+
